FinSmartBackend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
pandas
openpyxl
requests
langchain
langchain-core
langchain-community
langchain-groq
langchain-openai
langchain-astradb
langchain-huggingface
langchain-nvidia-ai-endpoints
sentence-transformers
torch
torchvision
torchaudio
crewai
crewai-tools
yfinance
